MainActivity menu changed. Location permission granting fixed. Profile picture load on NavView bug fixed.

+++ app/src/main/java/com/tim11/pma/ftn/pmaprojekat/DetailViewActivity.java
@@ -16,42 +16,30 @@ import android.support.v7.app.ActionBarDrawerToggle;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.view.menu.ActionMenuItemView;
 import android.support.v7.widget.Toolbar;
-import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-
-import com.tim11.pma.ftn.pmaprojekat.listener.RefreshNameDrawerListener;
-import com.tim11.pma.ftn.pmaprojekat.model.Hotel;
-import com.tim11.pma.ftn.pmaprojekat.model.User;
-import com.tim11.pma.ftn.pmaprojekat.util.PreferenceUtil;
-
 import org.androidannotations.annotations.Background;
-
+import org.androidannotations.annotations.EActivity;
 
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
-
-import com.j256.ormlite.android.apptools.OpenHelperManager;
-import com.j256.ormlite.dao.Dao;
-import com.tim11.pma.ftn.pmaprojekat.data.db.DatabaseHelper;
-import com.tim11.pma.ftn.pmaprojekat.data.db.HotelDAO;
-
-import com.tim11.pma.ftn.pmaprojekat.model.internal.HotelInternalModel;
-
-import org.androidannotations.annotations.Click;
-import org.androidannotations.annotations.EActivity;
-import org.androidannotations.annotations.ViewById;
-
 import java.sql.SQLException;
 
+import com.j256.ormlite.android.apptools.OpenHelperManager;
+import com.tim11.pma.ftn.pmaprojekat.data.db.DatabaseHelper;
+import com.tim11.pma.ftn.pmaprojekat.data.db.HotelDAO;
+import com.tim11.pma.ftn.pmaprojekat.listener.RefreshNameDrawerListener;
+import com.tim11.pma.ftn.pmaprojekat.model.internal.HotelInternalModel;
+import com.tim11.pma.ftn.pmaprojekat.model.Hotel;
+import com.tim11.pma.ftn.pmaprojekat.model.User;
+import com.tim11.pma.ftn.pmaprojekat.util.PreferenceUtil;
 
 import layout.HotelDetailsFragment_;
-import layout.HotelInformationFragment;
 
 @EActivity
 public class DetailViewActivity extends AppCompatActivity
@@ -167,21 +155,23 @@ public class DetailViewActivity extends AppCompatActivity
         if (id == R.id.action_favourites) {
 
             ActionMenuItemView menu = (ActionMenuItemView) findViewById(R.id.action_favourites);
-            DatabaseHelper helper = OpenHelperManager.getHelper(getApplicationContext(),DatabaseHelper.class);
+            DatabaseHelper helper = OpenHelperManager.getHelper(getApplicationContext(), DatabaseHelper.class);
             HotelDAO hotelDao = null;
             try {
                 hotelDao = helper.getHotelDao();
             } catch (SQLException e) {
                 e.printStackTrace();
             }
-            if(!isFavourite){
-                hotelDao.addToFavourites(hotel);
-                isFavourite = true;
-                menu.setIcon(getDrawable(R.drawable.ic_favourite_on));
-            }else{
-                hotelDao.removeFromFavourites(hotel);
-                menu.setIcon(getDrawable(R.drawable.ic_favourite_off));
-                isFavourite = false;
+            if (hotelDao != null) {
+                if (!isFavourite) {
+                    hotelDao.addToFavourites(hotel);
+                    isFavourite = true;
+                    menu.setIcon(getDrawable(R.drawable.ic_favourite_on));
+                } else {
+                    hotelDao.removeFromFavourites(hotel);
+                    menu.setIcon(getDrawable(R.drawable.ic_favourite_off));
+                    isFavourite = false;
+                }
             }
 
             return true;
